+++ README.md
@@ -2,7 +2,7 @@
 
 # 기능 목록
 ## step1
-- [ ] 영화 리스트 Activity 를 새로운 Fragment 로 바꾼다.
+- [x] 영화 리스트 Activity 를 새로운 Fragment 로 바꾼다.
 - [ ] Bottom Navigation 레아이웃을 작성한다.
 - [ ] 각 Navigation 의 Fragment 레이아웃을 작성한다.
 - [ ] Bottom Navigation 과 각 Fragment 를 연결해준다.
